Actualizacion de metodos y autorizacion por token

+++ src/main/java/com/porfolio/alecarb/entity/Educacion.java
@@ -18,9 +18,8 @@ public class Educacion {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    //@Column(name = "image_est")
-    //private String image_est;
+    private Integer id;
+   
     @Column(name = "institucion")
     private String institucion;
     @Column(name = "titulo")
@@ -29,8 +28,7 @@ public class Educacion {
     private String fecha_inicio;
     @Column(name = "Fecha_fin")
     private String fecha_fin;
-    //@Column (name = "en_curso")
-    //private boolean en_curso;
+    
     
     
     public Educacion(){
@@ -39,13 +37,13 @@ public class Educacion {
 
     public Educacion( String institucion, String titulo, String fecha_inicio, String fecha_fin, Persona persona) {
         
-        //this.image_est = image_est;
+       
         this.persona = persona;
         this.institucion = institucion;
         this.titulo = titulo;
         this.fecha_inicio = fecha_inicio;
         this.fecha_fin = fecha_fin;
-        //this.en_curso = en_curso;
+        
        
     }
