+++ src/main/java/com/porfolio/alecarb/repository/IEducacionRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IEducacionRepository extends JpaRepository<Educacion, Long>{
+public interface IEducacionRepository extends JpaRepository<Educacion, Integer>{
     
 }
