+++ src/main/java/com/porfolio/alecarb/repository/IHardSkillRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IHardSkillRepository extends JpaRepository<HardSkill, Long>{
+public interface IHardSkillRepository extends JpaRepository<HardSkill, Integer>{
     
 }
