+++ src/main/java/com/porfolio/alecarb/repository/INavbarRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface INavbarRepository extends JpaRepository<Navbar, Long>{
+public interface INavbarRepository extends JpaRepository<Navbar, Integer>{
     
 }
