+++ src/main/java/com/porfolio/alecarb/repository/ITrabajoRepository.java
@@ -6,6 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ITrabajoRepository extends JpaRepository<Trabajo, Long> {
+public interface ITrabajoRepository extends JpaRepository<Trabajo, Integer> {
+   
+    
     
 }
