+++ src/main/java/com/porfolio/alecarb/security/repository/IRolRepository.java
@@ -8,6 +8,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IRolRepository extends JpaRepository<Rol, Long>{
+public interface IRolRepository extends JpaRepository<Rol, Integer>{
     Optional<Rol> findByRolNombre(RolNombre rolNombre);
 }
